[3 - 4단계 방탈출 예약 대기] 커비(이현지) 미션 제출합니다. (#99)
* docs: 예약 대기 기능 구현 목록 추가

* refactor: 예약 조회에 fetch join 적용

* refactor(application.properties): 개행 추가

* test(ThemeRepositoryTest): TestFixture를 data.sql에서 메서드 내 주입 방식으로 변경

* refactor: 로그인 하지 않은 사용자의 예약 페이지 접근 권한 제한

* feat: 예약 대기 요청 기능 API 구현

* refactor: 예약 조회 시, 예약 대기 및 번호도 함께 조회하도록 기능 추가

* refactor: 중복 예약 대기 불가능하도록 반영

* refactor: 예약 대기 취소 기능 API 구현

* fix: 로그인 오류 수정

* docs: 예약 대기 관리 기능 구현 목록 추가

* feat: 관리자의 예약 대기 목록 조회 기능 API 구현

* feat: 관리자의 예약 대기 취소 기능 API 구현

* feat: 관리자의 예약 대기 취소 기능 API 구현

* feat: 자동 예약 대기 승인 기능 구현

* refactor: waiting 컨트롤러, 서비스 분리

* chore: ddl auto 설정 오타 수정

* refactor: 중복 예약 및 대기 제한 정책 반영

* refactor: reservationService에서 waiting 관련 로직 분리

* refactor: js 코드의 주석 제거

* refactor: WaitingController 삭제 및 통합, 메서드 네이밍 수정

* refactor: 불필요한 어노테이션 및 개행 삭제

* test: 예약 대기 관리 인수 테스트 추가

* test: 예약 대기 번호 및 자동 승인 관리 테스트 추가

* test(AcceptanceTest): test fixture 사용자 추가

* test: 예약 대기 관리 서비스 테스트 추가

* refactor: 주석 제거

* refactor: 예약 대기 취소 API의 URI 변경 및 WaitingController로 이동

* test:  test_schema.sql 삭제 및 인기 테마 조회를 위한 test fixture를 메서드 주입 방식으로 변경

* refactor: 불필요한 lazy loading 제거

* refactor: 전체 테이블 column에 not null 속성 추가

* refactor: 불필요한 검증 조건 제거

* refactor: 로그인 하지 않은 사용자의 내 예약 페이지 접근 권한 제한

* refactor: 개행 추가

* refactor: 불필요한 검증 조건 제거

* refactor: 관리자 접근 관련 status code 상수로 대체

* refactor: interceptor에 메서드 체이닝 적용

* refactor: 불필요한 생성자 제거

* refactor: 쿠키 및 토큰 유무 검증 방식 변경

* refactor(TokenProvider): 메서드 네이밍 변경

* test: displayName 추가

* test(ReservationWithWaitingTest): 테스트 추가

* refactor: 예약 항목 도메인 이름 변경

* refactor: Dto 클래스 네이밍 변경

* refactor(ReservationEntry): 메서드 분리

* refactor: 날짜, 시간, 테마를 포함하는 Schedule 도메인 생성 및 waiting에 적용

* refactor: Schedule VO Reservation에 적용

* refactor(ReservationEntry): 메서드명 변경

* refactor: 예약과 예약 대기 도메인의 중복제거를 위한 Base Entity 적용

* refactor: entity 패키지 분리

* refactor: Waiting 엔티티가 Reservation을 갖도록 정규화 적용

* refactor: Schedule 도메인 제거 적용

* refactor: 예약 대기의 예약 전환 로직 WaitingService로 분리

* refactor: 트랜잭션 적용

+++ src/main/java/roomescape/service/ReservationService.java
@@ -4,38 +4,48 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.Member;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Theme;
+import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.ReservationEntry;
+import roomescape.domain.WaitingWithRank;
+import roomescape.entity.Member;
+import roomescape.entity.Reservation;
+import roomescape.entity.ReservationTime;
+import roomescape.entity.Theme;
 import roomescape.handler.exception.CustomException;
 import roomescape.handler.exception.ExceptionCode;
 import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
+import roomescape.repository.WaitingRepository;
 import roomescape.service.dto.request.ReservationConditionRequest;
 import roomescape.service.dto.request.ReservationRequest;
-import roomescape.service.dto.response.MyReservationResponse;
+import roomescape.service.dto.response.MyReservationEntryResponse;
 import roomescape.service.dto.response.ReservationResponse;
 
 @Service
+@Transactional
 public class ReservationService {
 
     private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
     private final ThemeRepository themeRepository;
     private final MemberRepository memberRepository;
+    private final WaitingRepository waitingRepository;
+    private final WaitingService waitingService;
 
     public ReservationService(ReservationRepository reservationRepository,
                               ReservationTimeRepository reservationTimeRepository,
                               ThemeRepository themeRepository,
-                              MemberRepository memberRepository
+                              MemberRepository memberRepository, final WaitingRepository waitingRepository,
+                              final WaitingService waitingService
     ) {
         this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
         this.themeRepository = themeRepository;
         this.memberRepository = memberRepository;
+        this.waitingRepository = waitingRepository;
+        this.waitingService = waitingService;
     }
 
     public ReservationResponse createReservation(ReservationRequest reservationRequest, Long memberId) {
@@ -43,7 +53,7 @@ public class ReservationService {
         ReservationTime reservationTime = getReservationTime(reservationRequest.timeId());
         Theme theme = getTheme(reservationRequest.themeId());
 
-        if (reservationRepository.existsByTimeAndDate(reservationTime, reservationRequest.date())) {
+        if (reservationRepository.existsByDateAndTime(reservationRequest.date(), reservationTime)) {
             throw new CustomException(ExceptionCode.DUPLICATE_RESERVATION);
         }
         validateIsPastTime(reservationRequest.date(), reservationTime);
@@ -70,26 +80,35 @@ public class ReservationService {
     public List<ReservationResponse> findAllReservationsByCondition(ReservationConditionRequest condition) {
         Theme findTheme = getTheme(condition.themeId());
         Member findMember = getMember(condition.memberId());
-        List<Reservation> reservations = reservationRepository.findAllByThemeAndMemberAndDateBetween(
-                findTheme, findMember, condition.dateFrom(), condition.dateTo());
+        List<Reservation> reservations = reservationRepository.findAllByMemberAndThemeAndDateBetween(
+                findMember, findTheme,
+                condition.dateFrom(), condition.dateTo());
 
         return reservations.stream()
                 .map(ReservationResponse::from)
                 .toList();
     }
 
-    public List<MyReservationResponse> findAllByMemberId(Long id) {
-        Member member = memberRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
+    public List<MyReservationEntryResponse> findAllByMemberId(Long id) {
+        Member member = getMember(id);
         List<Reservation> reservations = reservationRepository.findAllByMember(member);
+        List<WaitingWithRank> waitingWithRanks = waitingService.findAllWithRankByMember(member);
 
-        return reservations.stream()
-                .map(MyReservationResponse::from)
+        List<ReservationEntry> reservationEntries = ReservationEntry.createEntries(reservations, waitingWithRanks);
+
+        return reservationEntries.stream()
+                .map(MyReservationEntryResponse::from)
                 .toList();
     }
 
     public void deleteReservation(Long id) {
-        reservationRepository.deleteById(id);
+        Reservation reservation = reservationRepository.findById(id)
+                .orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_RESERVATION));
+
+        if (!waitingRepository.existsByReservation(reservation)) {
+            reservationRepository.deleteById(id);
+        }
+        waitingService.convertWaitingToReservation(reservation);
     }
 
     private Theme getTheme(Long id) {
